refactor: abstract new punch screen logic to a custom hook
