+++ src/screens/new-punch/new-punch-screen.js
@@ -1,7 +1,6 @@
-import React, { useEffect } from "react";
+import React from "react";
 import Template from "../../components/template/template";
 import { moderateScale } from "react-native-size-matters";
-import { showToast, TOAST_POSITION, TOAST_TYPES } from "@utils/toast";
 import {
   Container,
   NewPunchDate,
@@ -16,62 +15,11 @@ import Button from "../../components/button/button";
 import SelectForm from "../../components/select/select-form";
 import InputForm from "../../components/input/input-form";
 import { FontAwesome5, MaterialCommunityIcons } from "@expo/vector-icons";
-import { format } from "date-fns";
-import { useForm } from "react-hook-form";
-import { yupResolver } from "@hookform/resolvers/yup";
-import * as yup from "yup";
-import { useNavigation } from "@react-navigation/native";
-
-const schema = yup
-  .object({
-    project: yup.string().required("Projeto obrigatório"),
-    morningFrom: yup
-      .string()
-      .required("Horário início de expediente obrigatório"),
-    morningTo: yup.string().required("Horário início de almoço obrigatório"),
-    afternoonFrom: yup.string().required("Horário final de almoço obrigatório"),
-    afternoonTo: yup
-      .string()
-      .required("Horário final de expediente obrigatório"),
-  })
-  .required();
+import useNewPunchScreen from "./use-new-punch-screen";
 
 const NewPunchScreen = () => {
-  const navigation = useNavigation();
-  const today = format(new Date(), "dd/MM/yyyy");
-
-  const {
-    handleSubmit,
-    formState: { errors },
-    control,
-  } = useForm({
-    defaultValues: {
-      project: "",
-      morningFrom: "",
-      morningTo: "",
-      afternoonFrom: "",
-      afternoonTo: "",
-    },
-    mode: "onBlur",
-    resolver: yupResolver(schema),
-  });
-
-  const onSubmit = (data) => {
-    console.log(data);
-  };
-
-  useEffect(() => {
-    const firstErrorKey = Object.keys(errors)[0];
-    const firstError = errors?.[firstErrorKey]?.message;
-
-    if (firstError) {
-      showToast({
-        text: firstError,
-        type: TOAST_TYPES.ERROR,
-        position: TOAST_POSITION.TOP,
-      });
-    }
-  }, [errors]);
+  const { control, handleSubmit, navigation, today, onSubmit } =
+    useNewPunchScreen();
 
   return (
     <Template>
